+++ src/config/mod.rs
@@ -1,13 +1,16 @@
 pub mod cargo;
 pub mod templating;
-
-use std::fs::File;
-use std::path::{Path, PathBuf};
+pub mod global;
 
 pub(crate) use crate::config::cargo::CargoConfig;
 use serde::{Deserialize, Serialize};
+use std::fs::File;
+use std::path::{Path, PathBuf};
+use std::str::FromStr;
 
-const GLOBAL_CONF_FILE: &'static Path = Path::new("global.toml");
+fn workspace_conf_file() -> PathBuf {
+    PathBuf::from_str("lion.toml").unwrap()
+}
 
 #[derive(Debug, Default, Serialize, Deserialize)]
 pub struct Config {
@@ -16,11 +19,11 @@ pub struct Config {
 
 impl Config {
     pub fn create(dir: impl AsRef<Path>) {
-        File::create(dir.as_ref().join(GLOBAL_CONF_FILE.to_path_buf())).unwrap();
+        File::create(dir.as_ref().join(workspace_conf_file())).unwrap();
     }
 
     pub fn load(dir: impl AsRef<Path>) -> Self {
-        let config_file = dir.as_ref().join(GLOBAL_CONF_FILE.to_path_buf());
+        let config_file = dir.as_ref().join(workspace_conf_file());
         if config_file.exists() {
             let config = std::fs::read_to_string(config_file).unwrap();
             toml::from_str(&config).unwrap()
